ws-dga-detection/requirements.txt
boto3==1.37.0
botocore==1.37.0
fastapi==0.115.12
huggingface_hub==0.25.2
keras==2.13.1
pydantic==2.10.6
python-dotenv==1.1.0
tensorflow==2.13.0
uvicorn==0.34.0
